macro impl seperated
